# g/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cashout": {
            "post": {
                "tags": [
                    "Transfert Moov"
                ],
                "summary": "Send money to a friend",
                "operationId": "7849595f510d162352b955b02b8a4047",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "msisdnThatReceive",
                                    "amount",
                                    "referenceid"
                                ],
                                "properties": {
                                    "msisdnThatReceive": {
                                        "description": "Recipient's identifier",
                                        "type": "string",
                                        "example": "63882818"
                                    },
                                    "amount": {
                                        "description": "Amount to send",
                                        "type": "string",
                                        "example": "100"
                                    },
                                    "referenceid": {
                                        "description": "Reference ID for the transfer",
                                        "type": "string",
                                        "example": "10000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Money sent successfully"
                    },
                    "400": {
                        "description": "Bad request, missing or invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided parameters are invalid or missing."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/get-balance": {
            "post": {
                "tags": [
                    "Transfert Moov"
                ],
                "summary": "Get user's mobile balance",
                "operationId": "784cd1016acf9defa140c08d1714e055",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "msisdn"
                                ],
                                "properties": {
                                    "msisdn": {
                                        "description": "User's mobile number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User's mobile balance retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request, missing or invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/mobile-status": {
            "post": {
                "tags": [
                    "Transfert Moov"
                ],
                "summary": "Get user's mobile status",
                "operationId": "ae3283d2bef12960e241d31519866517",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile"
                                ],
                                "properties": {
                                    "mobile": {
                                        "description": "User's mobile number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User's mobile status retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request, missing or invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/cash-in-transaction": {
            "post": {
                "tags": [
                    "Withdraw Moov"
                ],
                "summary": "Withdraw by ussd without pendding",
                "operationId": "65a3241c55931807819fd59ac3cc07af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "msisdn",
                                    "amount"
                                ],
                                "properties": {
                                    "msisdn": {
                                        "description": "The number on which the withdrawal is to be made",
                                        "type": "string"
                                    },
                                    "amount": {
                                        "description": "Amount to Withdraw",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Withdraw  with success"
                    },
                    "400": {
                        "description": "Bad request, missing or invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/get-cash-by-push": {
            "post": {
                "tags": [
                    "Withdraw Moov"
                ],
                "summary": "Withdraw by ussd with pendding",
                "operationId": "bee6cdb7f55d792d1d9e5ddb02293a83",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "msisdn",
                                    "amount"
                                ],
                                "properties": {
                                    "msisdn": {
                                        "description": "The number on which the withdrawal is to be made",
                                        "type": "string"
                                    },
                                    "amount": {
                                        "description": "Amount to Withdraw",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Withdraw  with success"
                    },
                    "400": {
                        "description": "Bad request, missing or invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
...
}